static-analysis/requirements.txt
python-magic
flask
networkx
python-dotenv
keystone-engine
loguru
lief
pexpect